+++ app/repositories/manager/__init__.py
@@ -1,0 +1,9 @@
+from .base import BaseManager
+from .beverage import BeverageManager
+from .index import IndexManager
+from .ingredient import IngredientManager
+from .order import OrderManager
+from .order_beverage import OrderBeverageManager
+from .order_ingredient import OrderIngredientManager
+from .report import ReportManager
+from .size import SizeManager
