+++ app/repositories/manager/base.py
@@ -1,0 +1,34 @@
+from typing import Any, Optional
+from app.plugins import ma, db
+
+
+class BaseManager:
+    model: Optional[db.Model] = None
+    serializer: Optional[ma.SQLAlchemyAutoSchema] = None
+    session = db.session
+
+    @classmethod
+    def get_all(cls):
+        serializer = cls.serializer(many=True)
+        _objects = cls.model.query.all()
+        result = serializer.dump(_objects)
+        return result
+
+    @classmethod
+    def get_by_id(cls, _id: Any):
+        entry = cls.model.query.get(_id)
+        return cls.serializer().dump(entry)
+
+    @classmethod
+    def create(cls, entry: dict):
+        serializer = cls.serializer()
+        new_entry = serializer.load(entry)
+        cls.session.add(new_entry)
+        cls.session.commit()
+        return serializer.dump(new_entry)
+
+    @classmethod
+    def update(cls, _id: Any, new_values: dict):
+        cls.session.query(cls.model).filter_by(_id=_id).update(new_values)
+        cls.session.commit()
+        return cls.get_by_id(_id)
