+++ app/repositories/models/order_beverage.py
@@ -1,0 +1,9 @@
+from app.plugins import db
+
+
+class OrderBeverage(db.Model):
+    _id = db.Column(db.Integer, primary_key=True)
+    beverage_price = db.Column(db.Float)
+    order_id = db.Column(db.Integer, db.ForeignKey('order._id'))
+    beverage_id = db.Column(db.Integer, db.ForeignKey('beverage._id'))
+    beverage = db.relationship('Beverage', backref=db.backref('beverage'))
