+++ app/repositories/serializer/order_beverage.py
@@ -1,0 +1,17 @@
+from app.plugins import ma
+
+from ..models import OrderBeverage
+from ..serializer import BeverageSerializer
+
+
+class OrderBeverageSerializer(ma.SQLAlchemyAutoSchema):
+
+    beverage = ma.Nested(BeverageSerializer)
+
+    class Meta:
+        model = OrderBeverage
+        load_instance = True
+        fields = (
+            'beverage_price',
+            'beverage'
+        )
