+++ app/repositories/serializer/size.py
@@ -1,0 +1,11 @@
+from app.plugins import ma
+
+from ..models import Size
+
+
+class SizeSerializer(ma.SQLAlchemyAutoSchema):
+
+    class Meta:
+        model = Size
+        load_instance = True
+        fields = ('_id', 'name', 'price')
